+++ src/model/films-model.js
@@ -5,10 +5,15 @@ const FILMS_NUMBER = 25;
 
 
 export default class FilmsModel {
-  films = Array.from({length: FILMS_NUMBER}, generateFilm);
-  comments = Array.from({length: 1000}, generateComment);
-  getFilms = () => this.films;
-  getComments = () => this.comments;
+  #films = Array.from({length: FILMS_NUMBER}, generateFilm);
+  #comments = Array.from({length: 1000}, generateComment);
 
+  get films() {
+    return this.#films;
+  }
+
+  get comments () {
+    return this.#comments;
+  }
 }
