+++ src/view/extras-view.js
@@ -3,18 +3,19 @@ import {createElement} from '../render.js';
 const createExtraTemplate = () => '<section class="films-list films-list--extra"></section';
 
 export default class ExtraView {
-  getTemplate () {
+  #element = null;
+  get template () {
     return createExtraTemplate ();
   }
 
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element () {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
-    return this.element;
+    return this.#element;
   }
 
   removeElement () {
-    this.element = null;
+    this.#element = null;
   }
 }
