+++ src/view/films-list-container-view.js
@@ -3,18 +3,19 @@ import {createElement} from '../render.js';
 const createFilmsListContainerTemplate = () => '<div class="films-list__container"></div>';
 
 export default class FilmsListContainerView {
-  getTemplate () {
+  #element = null;
+  get template () {
     return createFilmsListContainerTemplate ();
   }
 
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element () {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
-    return this.element;
+    return this.#element;
   }
 
   removeElement () {
-    this.element = null;
+    this.#element = null;
   }
 }
