+++ src/view/films-list-section-view.js
@@ -5,18 +5,19 @@ const createFilmsListSectionTemplate = () => `<section class="films-list">
 </section>`;
 
 export default class FilmsListSectionView {
-  getTemplate () {
+  #element = null;
+  get template () {
     return createFilmsListSectionTemplate ();
   }
 
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element () {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
-    return this.element;
+    return this.#element;
   }
 
   removeElement () {
-    this.element = null;
+    this.#element = null;
   }
 }
