+++ src/view/films-section-view.js
@@ -4,18 +4,20 @@ import {createElement} from '../render.js';
 const createFilmsSectionTemplate = () => '<section class="films"></section>';
 
 export default class FilmsSectionView {
-  getTemplate () {
+  #element = null;
+
+  get template () {
     return createFilmsSectionTemplate ();
   }
 
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
-    return this.element;
+    return this.#element;
   }
 
   removeElement () {
-    this.element = null;
+    this.#element = null;
   }
 }
