--- src/view/form-view.js
@@ -1,25 +1,0 @@
-import {createElement} from '../render.js';
-
-
-const createFormTemplate = () =>  (`<form class="film-details__inner" action="" method="get">
-  </form>`);
-
-
-export default class FormView {
-
-
-  getTemplate () {
-    return createFormTemplate ();
-  }
-
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-    return this.element;
-  }
-
-  removeElement () {
-    this.element = null;
-  }
-}
