+++ src/view/navigation-view.js
@@ -8,18 +8,19 @@ const createNavigationTemplate = () => (`<nav class="main-navigation">
 </nav>`);
 
 export default class NavigationView {
-  getTemplate () {
+  #element = null;
+  get template () {
     return createNavigationTemplate();
   }
 
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element () {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
-    return this.element;
+    return this.#element;
   }
 
   removeElement () {
-    this.element = null;
+    this.#element = null;
   }
 }
