+++ src/view/profile-view.js
@@ -6,19 +6,20 @@ const createProfileTemplate = () => (`<section class="header__profile profile">
 </section>`);
 
 export default class ProfileView {
-  getTemplate () {
+  #element = null;
+  get template () {
     return createProfileTemplate();
   }
 
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element () {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
-    return this.element;
+    return this.#element;
   }
 
   removeElement () {
-    this.element = null;
+    this.#element = null;
   }
 }
