+++ src/view/show-more-button-view.js
@@ -3,18 +3,19 @@ import {createElement} from '../render.js';
 const createShowMoreButtonTemplate = () => '<button class="films-list__show-more">Show more</button>';
 
 export default class ShowMoreButtonView {
-  getTemplate () {
+  #element = null;
+  get template () {
     return createShowMoreButtonTemplate ();
   }
 
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element () {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
-    return this.element;
+    return this.#element;
   }
 
   removeElement () {
-    this.element = null;
+    this.#element = null;
   }
 }
