+++ src/view/sort-view.js
@@ -7,18 +7,19 @@ const createSortTemplate = () => (`<ul class="sort">
 </ul>`);
 
 export default class SortView {
-  getTemplate () {
+  #element = null;
+  get template () {
     return createSortTemplate();
   }
 
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element () {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
-    return this.element;
+    return this.#element;
   }
 
   removeElement () {
-    this.element = null;
+    this.#element = null;
   }
 }
