+++ src/view/top-rated-heading-view.js
@@ -3,18 +3,19 @@ import {createElement} from '../render.js';
 const createHeadingTemplate = () => '<h2 class="films-list__title">Top rated</h2>';
 
 export default class TopHeadingView {
-  getTemplate () {
+  #element = null;
+  get template () {
     return createHeadingTemplate ();
   }
 
-  getElement () {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
+  get element () {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
     }
-    return this.element;
+    return this.#element;
   }
 
   removeElement () {
-    this.element = null;
+    this.#element = null;
   }
 }
